set up react native navigation
