+++ index.ios.js
@@ -1,0 +1,1 @@
+import App from './App'
