+++ index.js
@@ -1,5 +1,5 @@
-/** @format */
 
+import React from 'react';
 import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
